LayoutTests/storage/indexeddb/mozilla/versionchange-abort-expected.txt

Test IndexedDB's aborting setVersion

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
initialVersion = db.version;
request = db.setVersion('2')
Deleted all object stores.
objectStore = db.createObjectStore('foo');
PASS db.objectStoreNames.length is 1
index = objectStore.createIndex('bar', 'baz');
PASS objectStore.indexNames.length is 1
event.target.transaction.abort();
PASS db.version is initialVersion
PASS db.objectStoreNames.length is 0
PASS successfullyParsed is true

TEST COMPLETE
